WFLY-9814 Fix XTS subsystem on JDK9
