+++ xts/src/test/java/org/jboss/as/xts/XTSSubsystemTestCase.java
@@ -69,16 +69,6 @@ public class XTSSubsystemTestCase extends AbstractSubsystemBaseTest {
     }
 
     @Test
-    public void testBootEAP620() throws Exception {
-        testBoot1_1_0(ModelTestControllerVersion.EAP_6_2_0);
-    }
-
-    @Test
-    public void testBootEAP630() throws Exception {
-        testBoot1_1_0(ModelTestControllerVersion.EAP_6_3_0);
-    }
-
-    @Test
     public void testBootEAP640() throws Exception {
         testBoot1_1_0(ModelTestControllerVersion.EAP_6_4_0);
     }
